feat(packages/sui-consents): add hook
